Simplify load test analysis
-  keep check for average duration per speaker, remove checks for amplitude profile, frequency and sender matching - takes too long on test machines;
- uses amplitude modulation for the N first clients to prevent noise level rising for Load Tests.

+++ test/integration/IntegrationTest.h
@@ -98,7 +98,7 @@ struct IntegrationTest : public ::testing::Test
         std::vector<double> dominantFrequencies;
         std::vector<std::pair<uint64_t, double>> amplitudeProfile;
         size_t audioSsrcCount = 0;
-        std::map<double, size_t> receivedBytes;
+        std::map<size_t, size_t> receivedBytes;
     };
 
     IntegrationTest();
